+++ src/main/java/net/frozenblock/lib/config/frozenlib_config/FrozenLibConfig.java
@@ -20,6 +20,7 @@ package net.frozenblock.lib.config.frozenlib_config;
 import blue.endless.jankson.Comment;
 import java.util.List;
 import me.shedaniel.autoconfig.annotation.ConfigEntry;
+import net.fabricmc.loader.api.FabricLoader;
 import net.frozenblock.lib.FrozenSharedConstants;
 import net.frozenblock.lib.config.api.instance.Config;
 import net.frozenblock.lib.config.api.instance.json.JsonConfig;
@@ -51,13 +52,18 @@ public class FrozenLibConfig {
 			public void onSync(FrozenLibConfig syncInstance) {
 				var config = this.config();
 				USE_WIND_ON_NON_FROZEN_SERVERS = config.useWindOnNonFrozenServers;
+				IS_DEBUG = config.isDebug && FabricLoader.getInstance().isDevelopmentEnvironment();
 			}
 		}
 	);
 
 	public static volatile boolean USE_WIND_ON_NON_FROZEN_SERVERS = true;
+	public static volatile boolean IS_DEBUG = false;
 
 	@Comment("Mods may override any of these options, but the config file will not change.")
+
+	@EntrySyncData("isDebug")
+	public boolean isDebug = false;
 
 	@EntrySyncData(value = "useWindOnNonFrozenServers", behavior = SyncBehavior.UNSYNCABLE)
 	public boolean useWindOnNonFrozenServers = true;