+++ src/main/java/net/frozenblock/lib/wind/api/WindManager.java
@@ -28,9 +28,12 @@ import java.util.Optional;
 import java.util.function.Function;
 import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
 import net.frozenblock.lib.math.api.EasyNoiseSampler;
+import net.frozenblock.lib.networking.FrozenNetworking;
 import net.frozenblock.lib.wind.impl.WindManagerInterface;
 import net.frozenblock.lib.wind.impl.WindStorage;
+import net.frozenblock.lib.wind.impl.networking.WindAccessPacket;
 import net.frozenblock.lib.wind.impl.networking.WindDisturbancePacket;
 import net.frozenblock.lib.wind.impl.networking.WindSyncPacket;
 import net.minecraft.core.BlockPos;
@@ -291,6 +294,14 @@ public class WindManager {
 		double windX = Mth.lerp(disturbanceAmount, this.windX * windScale, windDisturbance.x * windDisturbanceScale) * scale;
 		double windY = Mth.lerp(disturbanceAmount, this.windY * windScale, windDisturbance.y * windDisturbanceScale) * scale;
 		double windZ = Mth.lerp(disturbanceAmount, this.windZ * windScale, windDisturbance.z * windDisturbanceScale) * scale;
+
+		if (FrozenLibConfig.IS_DEBUG) {
+			FrozenNetworking.sendPacketToAllPlayers(
+				this.level,
+				new WindAccessPacket(pos)
+			);
+		}
+
 		return new Vec3(
 			Mth.clamp(windX, -clamp, clamp),
 			Mth.clamp(windY, -clamp, clamp),