+++ src/main/java/net/frozenblock/lib/worldgen/biome/mixin/OverworldBiomeBuilderMixin.java
@@ -43,7 +43,9 @@ public class OverworldBiomeBuilderMixin {
 	)
 	public void frozenLib$injectFrozenBiomesToOverworld(Consumer<Pair<Climate.ParameterPoint, ResourceKey<Biome>>> consumer, CallbackInfo info) {
 		for (FrozenBiome frozenBiome : FrozenBiome.getFrozenBiomes()) {
-			frozenBiome.injectToOverworld(consumer);
+			if (frozenBiome.isEnabled()) {
+				frozenBiome.injectToOverworld(consumer);
+			}
 		}
 	}
 }