+++ CHANGELOG.md
@@ -5,5 +5,9 @@ Make sure to clear this after each release
 Put changelog here:
 
 -----------------
-- Prevented softlocks during worldgen.
-- Removed the `frozenlib:heavy_items` and `frozenlib:can_interact_while_crouching` tags and their functionalities as they've seen no usage.
+- Added a `disable` method to the `FrozenBiome` that prevents it from automatically being injected into worldgen.
+- The `windoverride` command has been revamped and renamed to `wind.`
+- Added the `EntityLootHelper` class, currently containing the `shouldSmeltLoot` method.
+- Added the `TintRegistryHelper` class, making simple registry of block and item tints quicker.
+- Significantly optimized how structure info is sent to the client in debug mode.
+- Added the new `BeaconEffectRegistry` class, which can be used to add a Mob Effect to any Beacon tier.